python/requirementsddd.txt
addict==2.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.14
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
fonttools==4.54.1
freetype-py==2.5.1
fsspec==2024.10.0
h11==0.14.0
h5py==3.11.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipydatawidgets==4.3.5
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
open3d==0.18.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycocotools==2.0.7
pyglet==1.5.29
pygltflib==1.16.2
Pygments==2.18.0
PyOpenGL==3.1.0
pyparsing==3.1.4
pyquaternion==0.9.9
pyrender==0.1.45
python-dateutil==2.9.0.post0
python-engineio==4.10.1
python-socketio==5.11.4
pythreejs==2.4.2
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rpds-py==0.20.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
simple-websocket==1.1.0
six==1.16.0
stack-data==0.6.3
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.0
traitlets==5.14.3
traittypes==0.2.1
trimesh==4.5.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.28
ultralytics-thop==2.0.10
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.16.0
wsproto==1.2.0
zipp==3.20.2
